Now it does not compile

+++ factorial.py
@@ -4,7 +4,7 @@ import numpy as np
 
 @njit(cache=True)
 def factorial(n):
-    r = np.float64(1)
+    r = np.float128(1)
     for i in range(1, n + 1):
         r *= i
     return r
